update get loan payments

+++ aiof.api.services/AiofMetadataRepository.cs
@@ -71,7 +71,7 @@ namespace aiof.api.services
             return await GetMetadataAsync("frequencies");
         }
 
-        public async Task<object> GetLoanPaymentsAsync(float loanAmount, float numberOfYears, float rateOfInterest, string frequency = "monthly")
+        public async Task<object> GetLoanPaymentsAsync(double loanAmount, double numberOfYears, double rateOfInterest, string frequency = "monthly")
         {
             _logger.LogInformation($"getting loan payments information. loanAmount='{loanAmount}', numberOfYears='{numberOfYears}', " +
                 $"rateOfInterest='{rateOfInterest}', frequency='{frequency}'");
